requirements.txt
pycryptodome
pyOpenSSL